SDK-1540: Move files into request package
