+++ docscan/sandbox/request/check/document_authenticity_check_test.go
@@ -4,11 +4,11 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/getyoti/yoti-go-sdk/v3/docscan/sandbox/check/report"
-	"github.com/getyoti/yoti-go-sdk/v3/docscan/sandbox/filter"
+	"github.com/getyoti/yoti-go-sdk/v3/docscan/sandbox/request/check/report"
+	"github.com/getyoti/yoti-go-sdk/v3/docscan/sandbox/request/filter"
 )
 
-func Example_documentTextDataCheckBuilder() {
+func Example_documentAuthenticityCheckBuilder() {
 	breakdown, err := report.NewBreakdownBuilder().
 		WithResult("some_result").
 		WithSubCheck("some_check").
@@ -29,11 +29,10 @@ func Example_documentTextDataCheckBuilder() {
 		return
 	}
 
-	check, err := NewDocumentTextDataCheckBuilder().
+	check, err := NewDocumentAuthenticityCheckBuilder().
 		WithBreakdown(breakdown).
 		WithRecommendation(recommendation).
 		WithDocumentFilter(filter).
-		WithDocumentField("some", "field").
 		Build()
 	if err != nil {
 		return
@@ -41,5 +40,5 @@ func Example_documentTextDataCheckBuilder() {
 
 	data, _ := json.Marshal(check)
 	fmt.Println(string(data))
-	// Output: {"result":{"report":{"recommendation":{"value":"some_value"},"breakdown":[{"sub_check":"some_check","result":"some_result","details":[]}]},"document_fields":{"some":"field"}},"document_filter":{"document_types":[],"country_codes":[]}}
+	// Output: {"result":{"report":{"recommendation":{"value":"some_value"},"breakdown":[{"sub_check":"some_check","result":"some_result","details":[]}]}},"document_filter":{"document_types":[],"country_codes":[]}}
 }
