+++ docscan/sandbox/request/check/document_check.go
@@ -1,7 +1,7 @@
 package check
 
 import (
-	"github.com/getyoti/yoti-go-sdk/v3/docscan/sandbox/filter"
+	"github.com/getyoti/yoti-go-sdk/v3/docscan/sandbox/request/filter"
 )
 
 type documentCheck struct {
