+++ docscan/sandbox/request/check/zoom_liveness_check.go
@@ -1,7 +1,7 @@
 package check
 
 import (
-	"github.com/getyoti/yoti-go-sdk/v3/docscan/sandbox/check/report"
+	"github.com/getyoti/yoti-go-sdk/v3/docscan/sandbox/request/check/report"
 )
 
 const (
