+++ docscan/sandbox/request/response_config_test.go
@@ -1,4 +1,4 @@
-package sandbox
+package request
 
 import (
 	"encoding/json"
